Allow audit logging to be configured when running in remote mode
RemoteRunner is the runner for a Yaybu process that is acting as
a resource provider and config pusher - the "master" end of a
push deployment.

Runner is the runner that will be involved in any actual changes
to the system.

So the new approach to setting up logging is to have audit logging
for the Runner and session logging for the RemoteRunner.