A better reimplementation of subprocess communicate
As with the python2.6 subprocess.Popen.communicate, the yaybu.core.shell
reimplementation of communicate waits for stdout and stderr handles to
disappear however the underlying process might have disappeared, when in
some instances the process may pass these handles to other, non-child
processes.

This patch provides an implementation closer to the python version 3
subprocess communicate function, which uses select.select with a
timeout - we've decided to set a one second timeout, when we break from
the select.select and run .poll() on the Popen object to see if the
command has a return code yet. If it has, the process has disappeared
and we should break from the communicate loop, when p.wait() picks up
the returncode as normal.